requirements.txt
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
fpdf>=1.7.2
PyYAML>=6.0
pathlib2>=2.3.7; python_version < "3.4" 